requirements.txt
amqp==5.2.0
annoy==1.17.3
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
dask==2024.9.1
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
faiss-cpu==1.8.0.post1
filelock==3.16.1
fsspec==2024.9.0
gunicorn==23.0.0
huggingface-hub==0.25.1
idna==3.10
ijson==3.3.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
locket==1.0.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.3
partd==1.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
toolz==0.12.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
